+++ android-slack/src/main/java/com/genymobile/android_slack/internal/ChannelsListSlackResponse.java
@@ -1,4 +1,4 @@
-package com.genymobile.android_slack.impl;
+package com.genymobile.android_slack.internal;
 
 import java.util.List;
