+++ android-slack/src/main/java/com/genymobile/android_slack/internal/RtmStartSlackResponse.java
@@ -1,4 +1,4 @@
-package com.genymobile.android_slack.impl;
+package com.genymobile.android_slack.internal;
 
 /**
  * Response to an rtm.start api call.
